requirements.txt
scipy
keras
pandas
numpy
Pillow
